update string kv; added get; other fixes

+++ chubby_server/src/constants.rs
@@ -1,4 +1,4 @@
 use std::time::Duration;
 
 pub const LEASE_EXTENSION: Duration = Duration::from_secs(12);
-pub const NUM_NODES: u32 = 5;
+pub const NUM_NODES: u64 = 5;
